refactor: extract core polling loop and completion logic from session_cmds_daemon_wait
- Extract handle_session_wait_with_hooks_and_sampler_output_mode (320 lines) to session_cmds_daemon_wait_core.rs
- Extract completion functions and exit code constants to session_cmds_daemon_wait_completion.rs
- Reduces main file from 611 to ~290 lines, below 8K token threshold
- Maintains all existing functionality and test compatibility
- Part of file splitting to resolve pre-commit monolith detection
