+++ crates/cli-sub-agent/src/gc_runtime_tests.rs
@@ -149,7 +149,7 @@ fn seed_runtime_session(
             artifacts: vec![SessionArtifact::new("output/summary.md")],
             peak_memory_mb: None,
             fallback_chain: None,
-        gate_timeout: false,
+            gate_timeout: false,
             manager_fields: Default::default(),
         },
     )
