+++ crates/cli-sub-agent/src/pipeline_tests_session_cleanup.rs
@@ -381,7 +381,7 @@ async fn state_dir_cap_failure_overwrites_stale_result_for_resume() {
             artifacts: Vec::new(),
             peak_memory_mb: None,
             fallback_chain: None,
-        gate_timeout: false,
+            gate_timeout: false,
             manager_fields: csa_session::SessionManagerFields {
                 report: Some(
                     toml::toml! {
