+++ crates/cli-sub-agent/src/session_cmds_result_tests.rs
@@ -394,7 +394,7 @@ fn build_result_json_payload_includes_result_sidecars() {
             artifacts: Vec::new(),
             peak_memory_mb: None,
             fallback_chain: None,
-        gate_timeout: false,
+            gate_timeout: false,
             manager_fields: Default::default(),
         },
         manager_sidecar: Some(
@@ -439,7 +439,7 @@ fn build_result_json_payload_redacts_result_sidecars() {
             artifacts: Vec::new(),
             peak_memory_mb: None,
             fallback_chain: None,
-        gate_timeout: false,
+            gate_timeout: false,
             manager_fields: Default::default(),
         },
         manager_sidecar: Some(
