+++ crates/cli-sub-agent/src/todo_errors_cmd.rs
@@ -237,7 +237,7 @@ mod tests {
             artifacts: Vec::new(),
             peak_memory_mb: None,
             fallback_chain: None,
-        gate_timeout: false,
+            gate_timeout: false,
             manager_fields: Default::default(),
         }
     }
