+++ crates/csa-session/src/manager_result_spillover.rs
@@ -332,7 +332,7 @@ mod tests {
             artifacts: vec![SessionArtifact::new("output/acp-events.jsonl")],
             peak_memory_mb: None,
             fallback_chain: None,
-        gate_timeout: false,
+            gate_timeout: false,
             manager_fields: crate::result::SessionManagerFields {
                 report: Some(
                     toml::toml! {
@@ -355,7 +355,7 @@ mod tests {
         let turn_2_result = SessionResult {
             summary: "turn 2".to_string(),
             fallback_chain: None,
-        gate_timeout: false,
+            gate_timeout: false,
             manager_fields: Default::default(),
             ..turn_1_result
         };
@@ -475,7 +475,7 @@ mod tests {
             artifacts: vec![SessionArtifact::new("output/acp-events.jsonl")],
             peak_memory_mb: None,
             fallback_chain: None,
-        gate_timeout: false,
+            gate_timeout: false,
             manager_fields: Default::default(),
         };
         save_result_in_with_threshold(
