+++ crates/csa-session/src/result.rs
@@ -241,7 +241,7 @@ mod tests {
             artifacts: vec![SessionArtifact::new("output/diff.patch")],
             peak_memory_mb: None,
             fallback_chain: None,
-        gate_timeout: false,
+            gate_timeout: false,
             manager_fields: Default::default(),
         };
 
@@ -274,7 +274,7 @@ mod tests {
             artifacts: vec![],
             peak_memory_mb: None,
             fallback_chain: None,
-        gate_timeout: false,
+            gate_timeout: false,
             manager_fields: Default::default(),
         };
 
@@ -361,7 +361,7 @@ artifacts = ["output/a.txt", "output/b.txt"]
             ],
             peak_memory_mb: None,
             fallback_chain: None,
-        gate_timeout: false,
+            gate_timeout: false,
             manager_fields: Default::default(),
         };
